+++ morpheus-graphql-core/src/Data/Morpheus/Types/Internal/AST.hs
@@ -150,6 +150,7 @@ module Data.Morpheus.Types.Internal.AST
     RootOperationTypeDefinition (..),
     UnionSelection,
     SchemaDefinition (..),
+    buildSchema,
   )
 where