+++ morpheus-graphql-client/changelog.md
@@ -4,6 +4,14 @@
 
 ### breaking changes
 
+- support of `schema`. issue #412
+
+  ```graphql
+  schema {
+    query: MyQuery
+  }
+  ```
+
 - from now you should provide for every custom graphql scalar definition coresponoding haskell type definition and `GQLScalar` implementation fot it. for details see [`examples-client`](https://github.com/morpheusgraphql/morpheus-graphql/tree/master/examples-client)
 
 - input fields and query arguments are imported without namespacing